refactor: share xai image size mapping
